refactor movie import, make studio nullable
