+++ MediaSet.Data/Movie/Movie.cs
@@ -11,7 +11,7 @@ namespace MediaSet.Data.MovieData
         public string Runtime { get; set; }
         public string Plot { get; set; }
         public string IMDBLink { get; set; }
-        public int StudioId { get; set; }
+        public int? StudioId { get; set; }
         public Studio Studio { get; set; }
         public ICollection<MovieDirector> MovieDirectors { get; set; }
         public ICollection<MovieProducer> MovieProducers { get; set; }
